fixed bugs in android - support file sizes over 2GB - handle double call of intent so destination file is not deleted - handle situation when app is closed as download is finished - bug when on progress is not returned when app starts

+++ android/src/main/java/com/eko/ProgressCallback.java
@@ -1,5 +1,5 @@
 package com.eko;
 
 public interface ProgressCallback {
-    void onProgress(String configId, int bytesDownloaded, int bytesTotal);
+    void onProgress(String configId, long bytesDownloaded, long bytesTotal);
 }
